applied requested changes
