+++ locales/en/index.yml
@@ -1,16 +1,17 @@
 ---
 idpay:
-  navigation:
-    header: Adesione all'iniziativa
-  pdndAcceptance:
-    requisites:
-      info:
-        header: Come avviene la verifica?
-        body: Comunichiamo direttamente con l’Ente che detiene l’informazione necessaria a effettuare il controllo, in questo caso **{{provider}}**.
-          In questo modo non devi produrre alcuna autocertificazione. Comodo, no?
-        understoodCTA: Ho capito
-    title: Possiedi i seguenti requisiti?
-    subtitle: Verificheremo con alcuni Enti se sei in possesso dei requisiti per poter richiedere {{service}}. 
+  onboarding:
+    navigation:
+      header: Adesione all'iniziativa
+    PDNDPrerequisites:
+      prerequisites:
+        info:
+          header: Come avviene la verifica?
+          body: Comunichiamo direttamente con l’Ente che detiene l’informazione necessaria a effettuare il controllo, in questo caso **{{provider}}**.
+            In questo modo non devi produrre alcuna autocertificazione. Comodo, no?
+          understoodCTA: Ho capito
+      title: Possiedi i seguenti requisiti?
+      subtitle: Verificheremo con alcuni Enti se sei in possesso dei requisiti per poter richiedere {{service}}. 
 date:
   formats:
     default: "%d/%m/%Y"
